% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mpc_srcs.R
\name{mpc_dtm_src}
\alias{mpc_dtm_src}
\title{Get the tile sources for Copernicus GLO-30, ALOS World 3D-30m, NASADEM HGT
v001 DTM using the planetary computer STAC catalogue.}
\usage{
mpc_dtm_src(
  aoi,
  collection = c("cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem")
)
}
\arguments{
\item{aoi}{an sf object to define where you want the data for.}

\item{collection}{Any of: "cop-dem-glo-30", "cop-dem-glo-90", "alos-dem", "nasadem".}
}
\value{

}
\description{
Get the tile sources for Copernicus GLO-30, ALOS World 3D-30m, NASADEM HGT
v001 DTM using the planetary computer STAC catalogue.
}
