% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstac_bbox.R
\name{rstac_bbox}
\alias{rstac_bbox}
\alias{rstac_bbox.sf}
\alias{rstac_bbox.sfc}
\alias{rstac_bbox.stars}
\alias{rstac_bbox.SpatRaster}
\alias{rstac_bbox.SpatVector}
\alias{rstac_bbox.numeric}
\title{Create an {rstac} compatible search boundary box}
\usage{
rstac_bbox(x)

\method{rstac_bbox}{sf}(x)

\method{rstac_bbox}{sfc}(x)

\method{rstac_bbox}{stars}(x)

\method{rstac_bbox}{SpatRaster}(x)

\method{rstac_bbox}{SpatVector}(x)

\method{rstac_bbox}{numeric}(x)
}
\arguments{
\item{x}{A numeric (length 4) to define the spatial bounds of the STAC API
query or one of the following rspatial objects: sf, sfc, stars, SpatVector,
SpatRaster}
}
\value{
A numeric of length 4 defining the xmin, ymin, xmax ymax coordinates
of the search extent in WGS84 (EPSG:4326).
}
\description{
Create an {rstac} compatible search boundary box
}
\examples{

mpc_dtm_src(c( -121.772, 45.321, -121.611, 45.419))

}
